requirements.txt
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
et-xmlfile==1.1.0
google-api-core==2.0.0
google-api-python-client==2.19.0
google-auth==2.0.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
mysql-connector-python==8.0.26
mysqlclient==2.0.3
oauth2client==4.1.3
openpyxl==3.0.7
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
PyDrive==1.3.1
pyparsing==2.4.7
python-dotenv==0.19.0
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
